requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tweepy==4.14.0
openai==1.3.7
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
schedule==1.2.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
jinja2==3.1.2
python-dateutil==2.8.2